feat: Allow solo hunting with Lone Wolf skill
Enables solo players to initiate hunting expeditions if they possess the 'lone-wolf' skill, by adjusting the party member check. This also simplifies the condition for starting a hunt to allow parties of 1 or more accepted members when the leader initiates.

The Dockerfile has been updated to use a multi-stage build, reducing the final image size.